# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Generated API",
    "description": "REST API generated by Backend Creator (bcm)",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "paths": {
...
    "/api/v1/furnitureBookings": {
      "get": {
        "tags": [
          "FurnitureBooking"
        ],
        "summary": "List all furnitureBookings",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20
            }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "order",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Search term for @bcm.searchable fields"
          },
          {
            "name": "include",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Comma-separated relation names to include"
          }
        ],
        "responses": {
          "200": {
            "description": "List of furnitureBookings",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/FurnitureBookingResponse"
                      }
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "page": {
                          "type": "integer"
                        },
                        "limit": {
                          "type": "integer"
                        },
                        "total": {
                          "type": "integer"
                        },
                        "totalPages": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "FurnitureBooking"
        ],
        "summary": "Create a new furnitureBooking",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FurnitureBookingCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "FurnitureBooking created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FurnitureBookingDataResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden — requires role: \"USER\", \"ADMIN\""
          },
          "422": {
            "description": "Validation error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/furnitureBookings/{id}": {
      "get": {
        "tags": [
          "FurnitureBooking"
        ],
        "summary": "Get furnitureBooking by key",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "include",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "FurnitureBooking details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FurnitureBookingDataResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "put": {
        "tags": [
          "FurnitureBooking"
        ],
        "summary": "Update furnitureBooking",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FurnitureBookingUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "FurnitureBooking updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FurnitureBookingDataResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden — requires role: \"USER\", \"ADMIN\""
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "patch": {
        "tags": [
          "FurnitureBooking"
        ],
        "summary": "Partially update furnitureBooking",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FurnitureBookingPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "FurnitureBooking updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FurnitureBookingDataResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden — requires role: \"USER\", \"ADMIN\""
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "FurnitureBooking"
        ],
        "summary": "Delete furnitureBooking",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden — requires role: \"USER\", \"ADMIN\""
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FurnitureBooking_FurnitureItemInput": {
        "type": "object",
        "properties": {
          "create": {
            "type": "object",
            "properties": {
              "title": {
                "type": "string"
              },
              "description": {
                "type": "string"
              },
              "price": {
                "type": "number",
                "format": "double"
              },
              "imageUrls": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "videoUrls": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "deletedAt": {
                "type": "string",
                "format": "date-time",
                "nullable": true
              }
            },
            "required": [
              "title",
              "description",
              "price"
            ],
            "description": "Create a related record"
          },
          "connect": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              }
            },
            "required": [
              "id"
            ],
            "description": "Connect to an existing record by unique selector"
          }
        },
        "description": "Nested input: provide either 'create' or 'connect'",
        "anyOf": [
          {
            "required": [
              "create"
            ]
          },
          {
            "required": [
              "connect"
            ]
          }
        ]
      },
      "FurnitureBooking_UserInput": {
        "type": "object",
        "properties": {
          "create": {
            "type": "object",
            "properties": {
              "name": {
                "type": "string"
              },
              "email": {
                "type": "string"
              },
              "password": {
                "type": "string"
              },
              "phone": {
                "type": "string",
                "nullable": true
              },
              "address": {
                "type": "string",
                "nullable": true
              },
              "rate": {
                "type": "number",
                "format": "float"
              },
              "role": {
                "$ref": "#/components/schemas/Role"
              },
              "isVerified": {
                "type": "boolean"
              }
            },
            "required": [
              "name",
              "email",
              "password"
            ],
            "description": "Create a related record"
          },
          "connect": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              }
            },
            "required": [
              "id"
            ],
            "description": "Connect to an existing record by unique selector"
          }
        },
        "description": "Nested input: provide either 'create' or 'connect'",
        "anyOf": [
          {
            "required": [
              "create"
            ]
          },
          {
            "required": [
              "connect"
            ]
          }
        ]
      },
      "FurnitureBookingResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "furnitureItemId": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "details": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/RequestStatus"
          },
          "adminNote": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "FurnitureBookingDataResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/FurnitureBookingResponse"
          }
        }
      },
      "FurnitureBookingCreate": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "details": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/RequestStatus"
          },
          "adminNote": {
            "type": "string",
            "nullable": true
          },
          "furnitureItem": {
            "$ref": "#/components/schemas/FurnitureBooking_FurnitureItemInput"
          },
          "user": {
            "$ref": "#/components/schemas/FurnitureBooking_UserInput"
          }
        },
        "required": [
          "name",
          "phone",
          "address",
          "furnitureItem",
          "user"
        ]
      },
      "FurnitureBookingUpdate": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "details": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/RequestStatus"
          },
          "adminNote": {
            "type": "string",
            "nullable": true
          },
          "furnitureItem": {
            "$ref": "#/components/schemas/FurnitureBooking_FurnitureItemInput"
          },
          "user": {
            "$ref": "#/components/schemas/FurnitureBooking_UserInput"
          }
        },
        "required": [
          "name",
          "phone",
          "address",
          "furnitureItem",
          "user"
        ]
      },
      "FurnitureBookingPatch": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "details": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/RequestStatus"
          },
          "adminNote": {
            "type": "string",
            "nullable": true
          },
          "furnitureItem": {
            "$ref": "#/components/schemas/FurnitureBooking_FurnitureItemInput"
          },
          "user": {
            "$ref": "#/components/schemas/FurnitureBooking_UserInput"
          }
        }
      },
...
      "Role": {
        "type": "string",
        "enum": [
          "USER",
          "ADMIN"
        ]
      },
      "RequestStatus": {
        "type": "string",
        "enum": [
          "PENDING",
          "ACCEPTED",
          "REJECTED",
          "CANCELLED",
          "AVAILABLE"
        ]
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}